updatepost test ok
