+++ backend/app/Http/Controllers/API/PostController.php
@@ -7,8 +7,8 @@ use App\Models\Post;
 use App\Models\Thread;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StorePost;
+use App\Http\Requests\UpdatePost;
 use App\Http\Resources\ThreadResources\PostResource;
- use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -21,8 +21,8 @@ class PostController extends Controller
     public function __construct()
     {
         $this->middleware('auth:api')->except(['index', 'show']);
-
         $this->middleware('filter_thread');
+
     }
 
     public function index($thread, Request $request)
@@ -64,7 +64,6 @@ class PostController extends Controller
         //return new PostResource($post);
         //应该要显示这个post，还有它的全部回帖，还有它的
 
-        $post = Post::where('id',$post->id);
         //return view('test', compact('posts'));
         //上面这一行代码，是为了通过debugler测试query实际效率。
         return response()->success([
@@ -90,19 +89,18 @@ class PostController extends Controller
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
-    public function update(StorePost $form,  $id)
+    public function update(UpdatePost $form, Thread $thread, Post $post)
     {
-        $post = Post::findorFail($id);
-        $thread=$post->thread;
-        $channel=$thread->channel;
-        if ((Auth::user()->canManageChannel($channel)==true)||((Auth::id() == $post->user_id)&&(!$thread->locked)&&($channel->allow_edit==true))){
+        try {
             $form->updatePost($post);
-            return response()->success([
-                'post' => new  PostResource($post),
-            ]);
-        }else{
+        }catch (Exception $e){
             return response()->error(config('error.403'), 403);
         }
+
+
+        return response()->success([
+            'post' => new  PostResource($post),
+        ]);
 
     }
