+++ backend/app/Http/Requests/StorePost.php
@@ -40,20 +40,7 @@ class StorePost extends FormRequest
         ];
     }
 
-    public function updatePost($post){
-
-        $data = $this->only('body');
-        $data['body'] = StringProcess::trimSpaces($data['body']);
-        $data['preview']=StringProcess::trimtext($data['body'], 50);
-        $data['is_anonymous']=$this->anonymous ? 1:0;
-        $data['use_markdown']=$this->markdown ? true:false;
-        $data['use_indentation']=$this->indentation ? true:false;
-        $data['allow_as_longpost']=$this->as_longcomment ? true:false;
-        $data['last_edited_at']=Carbon::now();
-        
-        $post->update($data);
-        return $post;
-    }
+  
     public function generatePost()
     {
         $thread = request()->route('thread');
